Move login from query to mutation and restructure project
