+++ 3_ecosystem/src/db/mod.rs
@@ -1,0 +1,2 @@
+pub mod model;
+pub mod repository;
