+++ 3_ecosystem/src/graphql/mod.rs
@@ -1,2 +1,2 @@
 pub mod handlers;
-pub(crate) mod model;
+pub mod model;
